feat!: use uint64_t instead of int for partition numbers

+++ examples/wordCount/wordcount/wordcount.cpp
@@ -61,7 +61,8 @@ private:
 };
 
 void run(const std::string &syncPrefix, const std::string &nodePrefix,
-         const std::string &subTopic, const std::vector<int> &topicPartitions) {
+         const std::string &subTopic,
+         const std::vector<uint64_t> &topicPartitions) {
   WordCounter compute;
   ndn::Face face;
 
@@ -101,7 +102,8 @@ int main(int argc, const char *argv[]) {
 
   std::string syncPrefix = config["syncPrefix"].as<std::string>();
   std::string nodePrefix = config["nodePrefix"].as<std::string>();
-  std::vector<int> partitions = config["partitions"].as<std::vector<int>>();
+  std::vector<uint64_t> partitions =
+      config["partitions"].as<std::vector<uint64_t>>();
   std::string subTopic = consumerConfig["topic"].as<std::string>();
 
   int saveInterval = measurementConfig["saveInterval"].as<int>();
